+++ .github/workflows/node.js.yml
@@ -45,7 +45,7 @@ jobs:
         run: npx prisma generate
 
       - run: npm run build --if-present
-      - run: (npm run web:nopdf &) | (timeout --preserve-status --foreground 10 cat; cat > /dev/null &) || exit 0
+      - run: (npm run web &) | (timeout --preserve-status --foreground 10 cat; cat > /dev/null &) || exit 0
         env:
           DATABASE_URL: "postgres://coronaschool:coronanervt@localhost:5432/coronaschool-test"
           ADMIN_AUTH_TOKEN: admintoken
