+++ integration-tests/WRITING_TESTS.md
@@ -4,7 +4,7 @@
 
 **Non Goals** of the tests is to cover each and every feature and ensure that nothing breaks, also we do not want to reach full coverage / test-driven development. We simply don't have capacity for that.
 
-To run integration tests locally, ensure that the environment variables `ADMIN_AUTH_TOKEN` and `INTEGRATION_TARGET` are set correctly in your `.env` (i.e. take them over from `.env.example`) and `SKIP_NOTIFICATION_IMPORT` is also set. Then start the Backend with `npm run web:nopdf` and run the integration tests with `npm run integration-tests:local`, or copy and paste these into a shell:
+To run integration tests locally, ensure that the environment variables `ADMIN_AUTH_TOKEN` and `INTEGRATION_TARGET` are set correctly in your `.env` (i.e. take them over from `.env.example`) and `SKIP_NOTIFICATION_IMPORT` is also set. Then start the Backend with `npm run web` and run the integration tests with `npm run integration-tests:local`, or copy and paste these into a shell:
 ```
 SKIP_NOTIFICATION_IMPORT=true ADMIN_AUTH_TOKEN=admin npm run web
 INTEGRATION_TARGET=http://localhost:5000/apollo ADMIN_AUTH_TOKEN=admin npm run integration-tests
