+++ package.json
@@ -18,7 +18,6 @@
     "web": "node -r source-map-support/register ./built/web/index.js",
     "web:dev": "node -r source-map-support/register -r dotenv/config ./built/web/index.js",
     "web:keep-db": "node -r source-map-support/register -r dotenv/config ./built/web/index.js --keepDB",
-    "web:nopdf": "node -r source-map-support/register -r dotenv/config ./built/web/index.js --noPDF",
     "web:debug": "npm run build && node --inspect=0 ./built/web/index.js",
     "web:docs": "apidoc -i web/controllers -i common -o web/public/docs",
     "typeorm": "npm run build:clean && npx typeorm",
@@ -67,7 +66,6 @@
     "graphql-scalars": "^1.10.0",
     "handlebars": "^4.7.7",
     "helmet": "^3.22.0",
-    "hpp": "^0.2.3",
     "html-pppdf": "^1.0.1",
     "keyv": "^4.0.3",
     "lodash": "^4.17.21",
